REF/TST: method-specific files for DataFrame timeseries methods (#32230)
