+++ pandas/tests/frame/conftest.py
@@ -1,8 +1,15 @@
+from itertools import product
+
 import numpy as np
 import pytest
 
 from pandas import DataFrame, NaT, date_range
 import pandas._testing as tm
+
+
+@pytest.fixture(params=product([True, False], [True, False]))
+def close_open_fixture(request):
+    return request.param
 
 
 @pytest.fixture
